layout: Move sizing logic from geom.rs to sizing.rs
Moves `Size`, `SizeConstraint`, `Sizes` and `LazySizeData`.
